requirements.txt
gradio~=6.1.0
pillow~=12.0.0